% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_tube.R
\name{bootstrap_curves}
\alias{bootstrap_curves}
\title{Bootstrap the curves via resampleing of each cluster}
\usage{
bootstrap_curves(
  dat,
  cluster_labels,
  lineages,
  cluster_group_list = NA,
  upscale_factor = NA,
  trials = 100,
  ncores = NA,
  verbose = F,
  ...
)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{lineages}{list of lineages (for example, estimated by the \code{slingshot} function)}

\item{cluster_group_list}{list denoting the hierarchy and order of the clusters}

\item{upscale_factor}{positive numeric (between 0 and 1) that controls how much to upweight the clusters,
with 1 being (almost) equal cluster sizes and 0 being no upweighting.}

\item{trials}{numeric}

\item{ncores}{numeric}

\item{verbose}{boolean}

\item{...}{additional parameters for \code{slingshot}}
}
\value{
a list
}
\description{
Bootstrap the curves via resampleing of each cluster
}
