% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{construct_3d_tube}
\alias{construct_3d_tube}
\title{Construct the 3D tube for each curve}
\usage{
construct_3d_tube(dat, radius, len = 20)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{radius}{a numeric}

\item{len}{a numeric}
}
\value{
a list
}
\description{
This function creates a circle roughly locally perpendicular to each point along a
curve, where the circle has radius \code{radius}. Here, \code{len} describes how many
points describe this circle, where in most applications, \code{len=20} is more than enough.
}
\details{
Importantly, \code{dat} needs to be in order, meaning the curve is represented by a discretization
of \code{nrow(dat)} points in 3-dimensional space (i.e., we require \code{ncol(dat) = 3}), whereby
the direct neighbors of any point (i.e., any row of \code{dat}) are the corresponding rows right before
and after it.
}
