% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_synonyms.R
\name{convert_synonyms}
\alias{convert_synonyms}
\title{Convert genes (in mouse) to a standardized set of synonyms}
\usage{
convert_synonyms(vec)
}
\arguments{
\item{vec}{a vector of gene names}
}
\value{
a vector of length \code{length(vec)}
}
\description{
This uses the database in \code{org.Mm.eg.db} package. At the time of running this
designing this code, we are using the version 3.10.0 released on	October 30, 2019, but
this code should work for even newer versions of \code{org.Mm.eg.db}.
}
\details{
If genes in \code{vec} are found to be in the database in \code{org.Mm.eg.db}, they are converted
to the their standardize symbol. If genes are not found in the database, they are left
untouched. Gene names in \code{vec} are replaced in-place.
}
