% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.absolute_threshold}
\alias{.absolute_threshold}
\title{Hard threshold all the entries in a matrix}
\usage{
.absolute_threshold(mat, direction, max_val = NA, tol2 = 0.001)
}
\arguments{
\item{mat}{numeric matrix with \code{n} rows and \code{p} columns}

\item{direction}{character either \code{"<="} or \code{">="} or \code{NA}}

\item{max_val}{maximum magnitude of each entry in the approximate}

\item{tol2}{numeric}
}
\value{
\code{n} by \code{d} matrix
}
\description{
While simple conceptually, this function is a bit janky-looking in its
implementation due to numeric instabilities previously encountered.
}
