% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.adaptive_gradient_step}
\alias{.adaptive_gradient_step}
\title{Adaptive projective gradient descent}
\usage{
.adaptive_gradient_step(dat, pred_mat, gradient_mat, k, max_val = NA,
  direction = "<=", stepsize_init = 100, stepdown_factor = 2,
  max_iter = 20, ...)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{pred_mat}{\code{n} by \code{d} matrix}

\item{gradient_mat}{\code{n} by \code{d} matrix}

\item{k}{numeric}

\item{max_val}{numeric or \code{NA}}

\item{direction}{"<=" or ">="}

\item{stepsize_init}{numeric}

\item{stepdown_factor}{numeric}

\item{max_iter}{numeric}

\item{...}{other parameters}
}
\value{
\code{n} by \code{d} matrix
}
\description{
The projective gradient descent is "adaptive" in the sense
that it will find an appropriate step size to ensure that after projection,
the objective value descreases. This is a heuristic to simply enable
reasonable results, not necessarily theoretically justified or computationally
efficient.
}
