% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{.circular_segmentation}
\alias{.circular_segmentation}
\title{Circular binary segmentation}
\usage{
.circular_segmentation(
  vec,
  resolution = 1/100,
  max_width_percentage = 0.1,
  hard_cut = NA
)
}
\arguments{
\item{vec}{the numeric vector to segment}

\item{resolution}{numeric value between 0 and 1 (exclusive)}

\item{max_width_percentage}{numeric value between 0 and 1 (exclusive), and must be larger than
\code{resolution}}

\item{hard_cut}{index between 1 and \code{length(vec)}, or \code{NA}}
}
\description{
Inspired by Olshen et al., 2004, this method does binary segmentation with a few modifications:
First, it only looks for changepoints, spaced out to be \code{resolution*length(vec)} apart. This
means it does not consider every index of \code{vec} to be a potential changepoint location.
Second, it only looks for starts and ends that are \code{max_width_percentage*length(vec)} apart.
This means it does not consider extremely long segments. Third, it has a parameter \code{hard_cut},
which is an index between \code{1} and \code{length(vec)}. Fourth, the metric used to segment
is not based on the difference on means. Rather, it is based on the difference between
the 25th quantile in the middle segment to the 75th quantile of the left and right shoulders.
Fifth (related to the fourth change), the change must be in a positive direction (i.e.,
the values in the middle segment must be higher than those in the shoulders, roughly speaking).
}
