% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_manipulation.R
\name{.compile_common_cells}
\alias{.compile_common_cells}
\title{Compile the common cells}
\usage{
.compile_common_cells(df_list)
}
\arguments{
\item{df_list}{output of \code{eSVD:::.extract_pseudotimes}}
}
\value{
a dataframe with four columns, \code{cell_idx}, \code{pseudotime},
\code{dist_to_curve} and \code{consensus}
}
\description{
Uses the output of \code{eSVD:::.extract_pseudotimes} (called \code{df_list})
and compiles all the information into
a data frame. Importantly, the maximum pseudotime in \code{df_list} across both data frames
divided by 10 is the tolerance, and if a cell's pseudotimes across both trajectories
differs more than than this tolerance, then its \code{consensus} in the output data frame
is \code{FALSE}. Additionally, when determining which pseudotime and distance-to-curve
to extract, this function uses the trajectory that the cell is closer to.
}
\details{
This function is hard-coded so \code{df_list} has to be length 2 (meaning there are two
estimated trajectories).
}
