% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_manipulation.R
\name{.compile_unique_cells}
\alias{.compile_unique_cells}
\title{Compile the unique cells}
\usage{
.compile_unique_cells(df_list)
}
\arguments{
\item{df_list}{output of \code{eSVD:::.extract_pseudotimes}}
}
\value{
a dataframe with three columns, \code{cell_idx}, \code{pseudotime},
\code{dist_to_curve}
}
\description{
Compared to \code{eSVD:::.compile_common_cells}, this function is a lot
simpler since there is nothing to compare between -- only extraction is needed
}
