% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.compute_cluster_distances}
\alias{.compute_cluster_distances}
\title{Compute distances between all the clusters}
\usage{
.compute_cluster_distances(dat, cluster_labels)
}
\arguments{
\item{dat}{matrix}

\item{cluster_labels}{vector with length equal to \code{nrow(dat)}}
}
\value{
matrix
}
\description{
Compute distances between all the clusters
}
