% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.compute_upscale_factor}
\alias{.compute_upscale_factor}
\title{Compute the upscale factor}
\usage{
.compute_upscale_factor(
  cluster_labels,
  cluster_intersection,
  upscale_factor = 0.5
)
}
\arguments{
\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{cluster_intersection}{list of vectors, where each vector contains entries in \code{cluster_labels}}

\item{upscale_factor}{numeric}
}
\value{
a numeric vector equal in length to \code{length(unique(cluster_labels))}
}
\description{
The upscale factor is the fraction of "maximum size" over "size of cluster", raised to \code{upscale_factor}.
Hence, if \code{upscale_factor=1}, all the resulting clusters in \code{cluster_intersection} are of the same size.
}
