% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.construct_3d_circle}
\alias{.construct_3d_circle}
\title{Construct a 3D circle}
\usage{
.construct_3d_circle(
  dat_vec,
  radius,
  basis_vec1,
  basis_vec2,
  len = 20,
  check = T
)
}
\arguments{
\item{dat_vec}{a vector of length 3}

\item{radius}{a positive numeric}

\item{basis_vec1}{a vector of length 3}

\item{basis_vec2}{a vector of length 3}

\item{len}{positive integer}

\item{check}{boolean}
}
\value{
a matrix with \code{len} rows and 3 columns
}
\description{
Given a 3-dimensional vector \code{dat_vec} and 2 basis vectors (also 3-dimensional)
\code{basis_vec1} and \code{basis_vec2}, output the discretization (of \code{len} points)
of a circle centered at \code{dat_vec} of radius \code{radius} that lies on the 2-dimensional
plane formed by \code{basis_vec1} and \code{basis_vec2}.
}
\details{
For this function, if \code{check = F}, we explicitly assume that \code{basis_vec1} and \code{basis_vec2}
are unit vectors that are perpendicular, without checking. Use this with caution
}
