% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.construct_all_circles}
\alias{.construct_all_circles}
\title{Construct all the circle for a given dataset}
\usage{
.construct_all_circles(dat, radius, len = 20)
}
\arguments{
\item{dat}{a numeric matrix with 3 columns}

\item{radius}{a positive numeric}

\item{len}{a positive integer}
}
\value{
a list of matrices, each with \code{len} rows and 3 columns
}
\description{
The dataset \code{dat} can only have 3 columns. This is essentially
a helper function to streamline the usage of \code{eSVD:::.find_adjacent_directions},
\code{eSVD:::.find_basis_vectors} and \code{eSVD:::.construct_3d_circle}.
}
