% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.construct_graph}
\alias{.construct_graph}
\title{Create an undirected graph from a distance matrix}
\usage{
.construct_graph(dist_mat)
}
\arguments{
\item{dist_mat}{(symmetric) distance matrix}
}
\value{
\code{igraph} object
}
\description{
Create an undirected graph from a distance matrix
}
