% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.construct_lineage_from_hierarchy}
\alias{.construct_lineage_from_hierarchy}
\title{Construct lineages from hiearchy}
\usage{
.construct_lineage_from_hierarchy(
  dist_mat,
  cluster_group_list,
  starting_cluster = 1
)
}
\arguments{
\item{dist_mat}{(symmetric) distance matrix}

\item{cluster_group_list}{list}

\item{starting_cluster}{numeric}
}
\value{
list, enumerating the different lineages
}
\description{
Using the order of clusters in \code{cluster_group_list}, construct the
lineage based on \code{dist_mat}. The function operates by appending clusters onto
existing trees. This operates in rounds equal to \code{length(cluster_group_list)}.
Specifically, within each round, the function first uses \code{eSVD:::.enumerate_dist_from_trees}
to populate a distance matrix from the previous round's trees, and then uses
\code{eSVD:::.enumerate_dist_between_levels} and \code{eSVD:::.enumerate_dist_within_levels} to
populate the new edges connecting the previous round's clusters to this round's clusters, as well
as this round's clusters to each other. Then, it constructs a list of trees via
\code{igraph::shortest_paths} and uses \code{eSVD:::.find_all_unique_paths} to "prune" the list
of trees by removing paths that are strictly contained in other paths. Then it proceeds to the next round.
}
