% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity.R
\name{.construct_neighborhood_graph}
\alias{.construct_neighborhood_graph}
\title{Construct nearest-neighbor graph}
\usage{
.construct_neighborhood_graph(mat, neighborhood_size)
}
\arguments{
\item{mat}{matrix with points represented as different rows}

\item{neighborhood_size}{positive integer}
}
\value{
\code{igraph} object
}
\description{
Construct nearest-neighbor graph
}
