% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.construct_resample_idx}
\alias{.construct_resample_idx}
\title{Construct resampled cluster labels}
\usage{
.construct_resample_idx(cluster_labels, upscale_vec)
}
\arguments{
\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{upscale_vec}{vector of positive values with length equal to \code{max(cluster_labels)}}
}
\value{
a vector of cluster labels (i.e. indices)
}
\description{
This function is designed so all the indices in 1 through \code{length(cluster_labels)} is
represented at least once. Typically, all the values in \code{upscale_vec} are 1 or larger.
}
