% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.covariance_distance}
\alias{.covariance_distance}
\title{Compute the Hotelling t test statistic for non-equal variances}
\usage{
.covariance_distance(
  mean_vec1,
  cov_mat1,
  n1,
  mean_vec2,
  cov_mat2,
  n2,
  tol = 1e-05
)
}
\arguments{
\item{mean_vec1}{vector}

\item{cov_mat1}{matrix}

\item{n1}{numeric}

\item{mean_vec2}{vector}

\item{cov_mat2}{matrix}

\item{n2}{numeric}

\item{tol}{numeric}
}
\value{
numeric
}
\description{
See \url{http://www.real-statistics.com/multivariate-statistics/hotellings-t-square-statistic/hotellings-t-square-unequal-covariance-matrices/}
}
