% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.enumerate_dist_between_levels}
\alias{.enumerate_dist_between_levels}
\title{Enumerate distances between paths and a set of indices}
\usage{
.enumerate_dist_between_levels(dist_mat, tree_list, cluster_vec)
}
\arguments{
\item{dist_mat}{(symmetric) distance matrix}

\item{tree_list}{list of paths, corresponding to indices from 1 to \code{nrow(dist_mat)}}

\item{cluster_vec}{list of indices from 1 to \code{nrow(dist_mat)}}
}
\value{
a matrix with 3 columns, representing the two indicies and the value from \code{dist_mat}
}
\description{
Enumerate edges corresponding to the last index in each element in \code{tree_list}
and any index in \code{cluster_vec}
}
