% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.enumerate_dist_from_trees}
\alias{.enumerate_dist_from_trees}
\title{Enumerate distances from paths stored in a list}
\usage{
.enumerate_dist_from_trees(dist_mat, tree_list)
}
\arguments{
\item{dist_mat}{(symmetric) distance matrix}

\item{tree_list}{list of paths, corresponding to indices with values between 1 and \code{nrow(dist_mat)}}
}
\value{
a matrix with 3 columns, representing the two indicies and the value from \code{dist_mat}
}
\description{
Enumerate distances from paths stored in a list
}
