% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.enumerate_dist_within_levels}
\alias{.enumerate_dist_within_levels}
\title{Enumerate distances within a vector of indices}
\usage{
.enumerate_dist_within_levels(dist_mat, cluster_vec)
}
\arguments{
\item{dist_mat}{(symmetric) distance matrix}

\item{cluster_vec}{list of indices from 1 to \code{nrow(dist_mat)}}
}
\value{
a matrix with 3 columns, representing the two indicies and the value from \code{dist_mat}
}
\description{
Enumerate distances within a vector of indices
}
