% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{.extract_information}
\alias{.extract_information}
\title{Extract segmentation information}
\usage{
.extract_information(segmentation_res)
}
\arguments{
\item{segmentation_res}{a list of outputs from \code{.find_highly_expressed_region}}
}
\value{
a data frame
}
\description{
This is used in \code{segment_genes_along_trajectories} to format the results
nicely.
}
