% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_manipulation.R
\name{.extract_pseudotimes}
\alias{.extract_pseudotimes}
\title{Extract pseudotimes}
\usage{
.extract_pseudotimes(slingshot_res)
}
\arguments{
\item{slingshot_res}{the output from the function \code{eSVD::slingshot}}
}
\value{
a list of dataframes, one for each trajectory in \code{slingshot_res}
}
\description{
Maps the (possibly) duplicated cells in \code{eSVD::slingshot} due to its
\code{upscale_factor} argument and extract the relevant information about each cell.
When a cell is detected to have multiple instances in \code{slingshot_res}, this functions
outputs its mean pseudotime as well as mean distance to curve among all its instances.
}
