% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.find_adjacent_directions}
\alias{.find_adjacent_directions}
\title{Find the local direction}
\usage{
.find_adjacent_directions(dat, idx)
}
\arguments{
\item{dat}{a numeric matrix}

\item{idx}{an index between 1 and \code{nrow(dat)}}
}
\value{
a unit vector
}
\description{
Find the local direction for row \code{idx} (an index) within a dataset \code{dat}.
This does it by averaging the unit vector between row \code{idx} and \code{idx+1},
and the unit vector between row \code{idx} and \code{idx-1} (if all three indices are valid).
Then, it normalizes this average again, so this function always returns a unit vector
(i.e., a vector with L2 norm equal to 1).
}
