% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.find_all_unique_paths}
\alias{.find_all_unique_paths}
\title{Remove duplicate paths}
\usage{
.find_all_unique_paths(path_list, starting_cluster)
}
\arguments{
\item{path_list}{list of vectors of indices}

\item{starting_cluster}{index}
}
\value{
a pruned version of \code{path_list}
}
\description{
We only keep paths in \code{path_list} that 1) start from \code{starting_cluster} and
2) are not strictly contained in any other path in our outputed list
}
