% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.find_basis_vectors}
\alias{.find_basis_vectors}
\title{Find basis vectors that are perpendicular to a vector}
\usage{
.find_basis_vectors(direction)
}
\arguments{
\item{direction}{a numeric vector of length 3}
}
\value{
a list of 2 unit vectors, both of length 3
}
\description{
Only works in 3-dimensional space (i.e., \code{length(direction) = 3}).
The \code{direction} vector need not be a unit vector. However, this function
is guaranteed to return two unit vectors.
}
