% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{.find_common_genes}
\alias{.find_common_genes}
\title{Find common genes}
\usage{
.find_common_genes(
  res_mat,
  traj_genes,
  common_n,
  threshold,
  number_of_genes = 2,
  manual_add = NA
)
}
\arguments{
\item{res_mat}{the 9-columned data frame that is the output of \code{segment_genes_along_trajectories}}

\item{traj_genes}{index of genes that are unique to either trajectories that should be removed, possibly left empty}

\item{common_n}{positive integer}

\item{threshold}{non-negative numeric}

\item{number_of_genes}{positive integer}

\item{manual_add}{vector of indices with values between \code{1} and \code{max(res_mat$idx)}
that are manually included, regardless of whether or not the gene's associated objective value
in \code{res_mat} exceeds \code{threshold}}
}
\value{
a vector of unique indices between \code{1} and \code{max(res_mat$idx)}, sorted
by pseudotime
}
\description{
Find genes based on \code{res_mat}, for genes in trajectory \code{traj} that
are below a pseudotime index of \code{common_n}, according to both segmentations (i.e.,
the \code{start_1} and \code{start_2} columns in \code{res_mat}), and then among
such that the both \code{obj_1} and \code{obj_2} columns in \code{res_mat} are above
threshold. These genes are then sorted according to \code{start_1} and \code{end_1}
(i.e., based on trajectory 1, for an arbitrary choice), so that the first \code{number_of_genes}
are selected at each pseudotime index based on their values in \code{obj_1}.
}
