% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.find_correct_orientation}
\alias{.find_correct_orientation}
\title{Align two circles}
\usage{
.find_correct_orientation(mat1, mat2)
}
\arguments{
\item{mat1}{a numeric matrix with 3 columns}

\item{mat2}{a numeric matrix with 3 columns and the \code{nrow(mat1)} number of rows.}
}
\value{
\code{mat2} with a new permutation of its rows
}
\description{
Given two circles in 3-dimensional space, represented as 2 different matrices with equal number of rows
and 3 columns, shuffle the row indices in \code{mat2} around so the rows of \code{mat2}
are (in some sense) aligned with the rows of \code{mat1}. This function has a rather
complicated-looking subroutine using \code{eSVD:::.subset_indices} since \code{mat1} could be
encoding a discretization of a circle in a clockwise direction while \code{mat2} could be
encoding a discretization of a circle in a counter-clockwise direction.
}
\details{
The code does not explicitly check that \code{mat1} and \code{mat2} contain points that
are essentially an ordered discretization of a circle, but it relies on this fact.
}
