% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{.find_highly_expressed_region}
\alias{.find_highly_expressed_region}
\title{Find highly specific genes - Pipeline}
\usage{
.find_highly_expressed_region(
  common_vec,
  specific_vec1,
  specific_vec2,
  standardize = T,
  ...
)
}
\arguments{
\item{common_vec}{a vector}

\item{specific_vec1}{a vector}

\item{specific_vec2}{a vector}

\item{standardize}{boolean}

\item{...}{extra arguments for \code{eSVD:::.circular_segmentation}}
}
\value{
a list
}
\description{
This function takes in three vectors, which is specialized in finding the
highly informative (i.e., highly expressive) genes between two trajectories.
It first uses \code{eSVD:::.np_smoother} to smooth the concatenation of \code{common_vec}
and \code{specific_vec1}, as well as to smooth the concatenation of \code{common_vec}
and \code{specific_vec2}, via \code{np::npreg} function (i.e., kernel regression).
Then, it standardizes these smoothed outputs jointly (so the joint vector has
mean 0 and standard deviation 1) if \code{standardize=T}. Finally, it separates
the standardedized joint vector back into two vectors and
\code{eSVD:::.circular_segmentation} to segment each of the vector. The only caveat is that
this separation keeps the smooths counterparts of \code{common_vec} and \code{specific_vec1}
and \code{specific_vec2} together, just in different orders. The reason for doing this
is because \code{eSVD:::.circular_segmentation} will have an argument \code{hard_cut},
which means to not search for a segmentation index past a certain index. This means
even when we're segmenting the first trajectory, we can retain the values from the second
trajectory for reference.
}
