% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{.find_trajectory_genes}
\alias{.find_trajectory_genes}
\title{Find trajectory genes}
\usage{
.find_trajectory_genes(
  res_mat,
  traj,
  common_n,
  n,
  threshold,
  number_of_genes = 2,
  manual_add = NA
)
}
\arguments{
\item{res_mat}{the 9-columned data frame that is the output of \code{segment_genes_along_trajectories}}

\item{traj}{the value \code{1} or \code{2}}

\item{common_n}{positive integer}

\item{n}{positive integer, larger than \code{common_n}}

\item{threshold}{non-negative numeric}

\item{number_of_genes}{positive integer}

\item{manual_add}{vector of indices with values between \code{1} and \code{max(res_mat$idx)}
that are manually included, regardless of whether or not the gene's associated objective value
in \code{res_mat} exceeds \code{threshold}}
}
\value{
a vector of unique indices between \code{1} and \code{max(res_mat$idx)}, sorted
by pseudotime
}
\description{
Find genes based on \code{res_mat}, for genes in trajectory \code{traj} that
are above a pseudotime index of \code{common_n} and below a pseudotime index of \code{n}.
These genes are first selected to be associated with an objective value larger than
\code{threshold} (according to \code{res_mat}) in the desired trajectory \code{traj}
as well as being lower than \code{threshold} in the other trajectory,
 and if any exist, the first \code{number_of_genes}
are considered at each pseudotime index.
}
\details{
(Note: It's possible that this function could've coded without the need for the \code{n} argument...)
}
