% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.fix_rank_defficiency_initialization}
\alias{.fix_rank_defficiency_initialization}
\title{Fix rank defficiency among two matrices}
\usage{
.fix_rank_defficiency_initialization(u_mat, v_mat, direction)
}
\arguments{
\item{u_mat}{a numeric matrix}

\item{v_mat}{a numeric matrix with the same number of columns as \code{u_mat}}

\item{direction}{character either \code{"<="} or \code{">="} or \code{NA}}
}
\value{
a list of \code{u_mat} and \code{v_mat}
}
\description{
Given two matrices, \code{u_mat} and \code{v_mat} (both with the same number of columns),
adjust these two matrices so the \code{u_mat \%*\% t(v_mat)} is actually of the desired
rank. This function is needed since sometimes upstream, the matrices \code{u_mat} and \code{v_mat}
do not actually have the rank equal to the number of columns (i.e., empirically we have
observed that \code{u_mat} might have a column that is all constant).
}
