% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.flatten_list}
\alias{.flatten_list}
\title{Flatten a list of lists}
\usage{
.flatten_list(x)
}
\arguments{
\item{x}{a list of lists}
}
\value{
a list
}
\description{
Code from \url{https://stackoverflow.com/questions/8139677/how-to-flatten-a-list-to-a-list-without-coercion/8139959#8139959}
}
