% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{.form_prediction_sbm}
\alias{.form_prediction_sbm}
\title{Form biclustering predictions based on clustering of rows and columns}
\usage{
.form_prediction_sbm(dat, row_clustering, col_clustering)
}
\arguments{
\item{dat}{a \code{n} by \code{p} matrix}

\item{row_clustering}{a clustering of \code{n} elements}

\item{col_clustering}{a clustering of \code{p} elements}
}
\value{
a \code{k} by \code{k} matrix
}
\description{
It is assumed that both \code{row_clustering} and \code{col_clustering} have the
same number of clusters
}
