% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.get_curves}
\alias{.get_curves}
\title{Estimate the slingshot curves}
\usage{
.get_curves(
  dat,
  cluster_labels,
  lineages,
  shrink = 1,
  stretch = 9999,
  thresh = 0.001,
  max_iter = 15,
  verbose = F
)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{lineages}{output of \code{.get_lineage()}}

\item{shrink}{shrinkage factor}

\item{stretch}{stretch factor}

\item{thresh}{parameter to determine convergence}

\item{max_iter}{maximum number of iterations}

\item{verbose}{boolean}
}
\value{
a list of \code{principal_curve} objects
}
\description{
Estimate the slingshot curves
}
