% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.initial_curve_fit}
\alias{.initial_curve_fit}
\title{Initial curve fit}
\usage{
.initial_curve_fit(lineages, cluster_vec, centers)
}
\arguments{
\item{lineages}{list output of \code{.get_lineages()}}

\item{cluster_vec}{a vector from 1 to \code{k}}

\item{centers}{matrix output of \code{compute_cluster_center()} that's
\code{k} by \code{d}}
}
\value{
a list of points
}
\description{
Initial curve fit
}
