% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.intersect_lineages_cluster_group_list}
\alias{.intersect_lineages_cluster_group_list}
\title{Intersect lineages with cluster group list}
\usage{
.intersect_lineages_cluster_group_list(lineages, cluster_group_list)
}
\arguments{
\item{lineages}{a list of vectors}

\item{cluster_group_list}{a list of vectors}
}
\value{
a list of vectors
}
\description{
Recall: \code{lineages} is a list of vectors, where each vector represents an ordered
set of indices. \code{cluster_group_list} is an ordered list of vectors, where each vector represent
indices "of the same tier." This function intersects the two by finding a list of vectors where
each of the output set of vectors is a self-contained vector in both \code{lineages} and \code{cluster_group_list}.
}
