% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{.np_smoother}
\alias{.np_smoother}
\title{Smoother function based on kernel regression}
\usage{
.np_smoother(vec)
}
\arguments{
\item{vec}{a vector that will be smoothed}
}
\value{
a vector of the same length as \code{vec}
}
\description{
The bandwidth selection is done automatically via \code{np::npregbw}.
}
