% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_factorization.R
\name{.optimize_mat}
\alias{.optimize_mat}
\title{Optimize a matrix, row-by-row}
\usage{
.optimize_mat(
  dat,
  current_mat,
  other_mat,
  left = T,
  max_val = NA,
  parallelized = F,
  verbose = F,
  ...
)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{p} columns represent genes}

\item{current_mat}{a matrix with \code{k} columns, and number of rows equal to either
\code{n} or \code{p}}

\item{other_mat}{a matrix with \code{k} columns, and number of rows equal to either
\code{n} or \code{p} (whichever one isn't what \code{current_mat} has)}

\item{left}{boolean}

\item{max_val}{maximum magnitude of the inner product}

\item{parallelized}{boolean}

\item{verbose}{boolean}

\item{...}{extra arguments, such as nuisance parameters for \code{"neg_binom"}
or \code{"curved gaussian"} for \code{family}}
}
\value{
matrix of the same dimension as \code{current_mat}
}
\description{
Given a data matrix \code{dat} where \code{attr(dat, "family")} is
\code{family} (i.e., tells us which negative
log-likelihood function we want to optimize over), optimize over each row of
\code{current_mat} holding the values in \code{other_mat} fixed. This function
determines which matrix is aligned with the rows or columns based on the setting of
\code{left}. If \code{left=TRUE}, then the rows of \code{current_mat}
correspond to the rows of \code{dat}, and if \code{left=FALSE}, then the
rows of \code{current_mat} correspond to the columns of \code{dat}.
}
\details{
This function does a gradient descent for every row in \code{current_mat}
via the function \code{eSVD:::.optimize_row}. See the documentation for that function for
more details.

If \code{parallelized=TRUE}, it is assumed the cores is already registered
via \code{doMC::registerDoMC}.
}
