% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.percent_shrinkage}
\alias{.percent_shrinkage}
\title{Determine the percentage shrinkage (a non-decreasing function)}
\usage{
.percent_shrinkage(pcurve, common_idx)
}
\arguments{
\item{pcurve}{output of \code{princurve::project_to_curve()}}

\item{common_idx}{indices to use}
}
\value{
vector
}
\description{
Determine the lambda's to use for interpolation based on the IQR for
lambda at \code{common_idx} indicies. Then use a linear interpolation
via \code{approx} to assign all the lambdas (even those not in \code{common_idx})
a relevant value based on the CDF of the cosine kernel.
}
