% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.project_rank_feasibility}
\alias{.project_rank_feasibility}
\title{Find low-rank and half-space approximate of a matrix}
\usage{
.project_rank_feasibility(
  mat,
  k,
  direction,
  max_val = NA,
  max_iter = 20,
  tol = 1e-06
)
}
\arguments{
\item{mat}{numeric matrix with \code{n} rows and \code{p} columns}

\item{k}{positive integer less than \code{min(c(n,p))}.}

\item{direction}{character either \code{"<="} or \code{">="} or \code{NA}}

\item{max_val}{maximum magnitude of each entry in the approximate}

\item{max_iter}{numeric}

\item{tol}{numeric}
}
\value{
\code{n} by \code{d} matrix
}
\description{
Alternating projection heuristic to find the approximate matrix of \code{mat}
that both is rank-\code{k} and has all its entries in the halfspace dictated
by \code{direction}. This function alternates
between using \code{eSVD:::.svd_projection} and \code{eSVD:::.absolute_threshold}.
}
\details{
However, if this function does not converge to a solution after \code{max_iter} iterations,
it gives up and uses \code{eSVD:::.sbm_projection} instead, which approximates \code{mat}
using a stochastic-block-model variant.
}
