% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.projection_matrix}
\alias{.projection_matrix}
\title{Compute the projection}
\usage{
.projection_matrix(vec, mat)
}
\arguments{
\item{vec}{a numeric vector}

\item{mat}{a numeric matrix with number of rows equal to \code{length(vec)} and with less
columns than rows}
}
\value{
a numeric vector of length \code{ncol(mat)}
}
\description{
Computes the projection of a vector \code{vec} onto the column space of \code{mat} (i.e.,
yields a vector with less dimensions than \code{length(vec)})
}
