% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.refine_curve_fit}
\alias{.refine_curve_fit}
\title{Initial fit of the curves}
\usage{
.refine_curve_fit(dat, curve_list, lineages, W, cluster_mat, stretch = 9999)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{curve_list}{sequence of points, input for \code{princurve::project_to_curve()}}

\item{lineages}{list output of \code{.get_lineages()}}

\item{W}{a weight matrix that is \code{n} by \code{num_lineage} (number of lineages)}

\item{cluster_mat}{0-1 matrix output of \code{.construct_cluster_matrix()} that
has \code{n} rows and \code{k} column}

\item{stretch}{stretch factor}
}
\value{
a list that contains the \code{num_lineage} curves as \code{principal_curve}
and a distance matrix (\code{D}) that contains the squared distance of each point
to its repsective lineage curve
}
\description{
This method is heavily dependent on \code{princurve::project_to_curve()}.
}
