% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.remove_duplicate_rows}
\alias{.remove_duplicate_rows}
\title{Remove duplicated rows}
\usage{
.remove_duplicate_rows(dat, tol = 1e-06)
}
\arguments{
\item{dat}{a numeric matrix}

\item{tol}{a small positive number}
}
\value{
a matrix
}
\description{
We remove rows that are both too similar to its immediate next row, as well as
rows that are too similar to the row after the immediate next row, both according to
the value in \code{tol}.
}
