% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{.reparameterize}
\alias{.reparameterize}
\title{Function to reparameterize two matrices}
\usage{
.reparameterize(u_mat, v_mat, equal_covariance = F)
}
\arguments{
\item{u_mat}{matrix of dimension \code{n} by \code{k}}

\item{v_mat}{matrix of dimension \code{p} by \code{k}}

\item{equal_covariance}{boolean}
}
\value{
list of two matrices
}
\description{

}
