% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.resample_all}
\alias{.resample_all}
\title{Resampling rows in a dattaset according to the amount of upscaling}
\usage{
.resample_all(
  dat,
  cluster_labels,
  cluster_group_list,
  lineages,
  upscale_factor
)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{cluster_group_list}{list denoting the hierarchy and order of the clusters}

\item{lineages}{list of vectors containing cluster labels}

\item{upscale_factor}{numeric scalar}
}
\value{
a list containing a new version of \code{dat} and \code{cluster_labels} but also
\code{idx_all} which stores the row-indices of the original \code{dat} used
}
\description{
This is the primary function used in \code{eSVD::slingshot} to decide which
rows in \code{dat} to resample (primarily useful for getting better-looking curves
in \code{eSVD:::.get_curves})
}
