% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{.sbm_projection}
\alias{.sbm_projection}
\title{Projection based on the stochastic block model}
\usage{
.sbm_projection(dat, k)
}
\arguments{
\item{dat}{matrix}

\item{k}{integer}
}
\value{
matrix of the same dimensions as \code{dat}, but has only at most \code{k^2}
unique values and is guaranteed to be rank \code{k}
}
\description{
Projection based on the stochastic block model
}
