% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.scale_vector}
\alias{.scale_vector}
\title{Scale a vector to lie between two endpoints}
\usage{
.scale_vector(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{vector}

\item{lower}{numeric}

\item{upper}{numeric}
}
\value{
vector
}
\description{
Takes \code{x} and scales it so the relative distances between points
are perseved and it ranges between \code{lower} and \code{upper}
}
