% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.shrink_to_avg}
\alias{.shrink_to_avg}
\title{Shrink curve to average}
\usage{
.shrink_to_avg(pcurve, avg_curve, pct, dat)
}
\arguments{
\item{pcurve}{output of \code{princurve::project_to_curve()}}

\item{avg_curve}{the average curve constructed by one of the elements
in the output of \code{.construct_average_curve()}}

\item{pct}{one of the elements in the output of \code{.percent_shrinkage()}}

\item{dat}{a \code{n} by \code{d} matrix}
}
\value{
a \code{principal_curve} object
}
\description{
Shrink curve to average
}
