% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.smoother_func}
\alias{.smoother_func}
\title{Smooth approximation based on lambdas}
\usage{
.smoother_func(lambda, dat, verbose = F)
}
\arguments{
\item{lambda}{vector given by one component of the \code{principal_curve} object}

\item{dat}{a \code{n} by \code{d} matrix. Here, \code{n} could be a subset of the
original dataset}

\item{verbose}{boolean}
}
\value{
a smoothed \code{n} by \code{d} matrix that is ordered by \code{lambda}
}
\description{
Does nonparametric regression (local polynomial regression)
of \code{dat} column-by-column based on the ordering given
by \code{lambda}.
}
\details{
This function re-orders the rows of \code{dat} according to \code{lambda}. While it's
not necessary for this function, other functions that depend on this function's output
will require this.
}
