% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{.spectral_clustering}
\alias{.spectral_clustering}
\title{Spectral clustering}
\usage{
.spectral_clustering(dat, k)
}
\arguments{
\item{dat}{matrix where spectral clustering will be applied to both the rows and columns}

\item{k}{integer}
}
\value{
a list \code{row_clustering} and \code{col_clustering} for the respective clusterings
}
\description{
Spectral clustering
}
