% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{.subset_indices}
\alias{.subset_indices}
\title{Construct a mapping from one set of indices to another}
\usage{
.subset_indices(n, anchor1, anchor2, direction_forward = T)
}
\arguments{
\item{n}{a positive integer}

\item{anchor1}{a positive integer less than \code{n}}

\item{anchor2}{a positive integer less than \code{n}}

\item{direction_forward}{a boolean}
}
\value{
a list of two vectors, each of length \code{n}, named \code{to} and \code{from}.
}
\description{
Given an index \code{n} and two indices \code{anchor1} and \code{anchor2} that are smaller
than \code{n}, output a vector \code{to} and \code{from} (as a list) where
the first element in \code{to} is \code{anchor1} and the first element of \code{from}
is \code{anchor2}. Then, if \code{direction_forward=TRUE}, while the indicies of
\code{to} are counting up, the indicies of \code{from}
are also counting up. Otherwise (meaning \code{direction_forward=FALSE}),
while the indicies of
\code{to} are counting up, the indicies of \code{from}
are also counting down.
}
\details{
In either case, all the indices here are (roughly) modulo \code{n}.
}
