% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_segmentation.R
\name{find_highly_expressed_region}
\alias{find_highly_expressed_region}
\title{Find highly expressed regions in a vector}
\usage{
find_highly_expressed_region(vec, resolution = 50)
}
\arguments{
\item{vec}{numeric vector}

\item{resolution}{value less than \code{length(vec)}}
}
\value{
numeric vector
}
\description{
The input vector \code{vec} should already be meaningfully ordered
}
