% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization}
\alias{initialization}
\title{Initialization for matrix factorization}
\usage{
initialization(
  dat,
  k = 2,
  family,
  max_val = NA,
  max_iter = 10,
  tol = 0.001,
  verbose = F,
  ...
)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{k}{positive integer less than \code{min(c(nrow(dat), ncol(dat)))}}

\item{family}{character (\code{"gaussian"}, \code{"exponential"}, \code{"poisson"}, \code{"neg_binom"},
or \code{"curved gaussian"})}

\item{max_val}{maximum magnitude of the inner product}

\item{max_iter}{numeric}

\item{tol}{numeric}

\item{verbose}{boolean}

\item{...}{extra arguments, such as nuisance parameters for \code{"neg_binom"}
or \code{"curved gaussian"} for \code{family}}
}
\value{
list containing \code{u_mat} and \code{v_mat}
}
\description{
This function uses \code{eSVD:::.projected_gradient_descent} primarily, so see
that function's documentation for more details.
}
