% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{order_highly_expressed_genes}
\alias{order_highly_expressed_genes}
\title{Ordering the highly expressed genes}
\usage{
order_highly_expressed_genes(
  res_mat,
  nrow1,
  nrow2,
  common_n,
  threshold,
  number_of_genes = 2,
  manual_add_common = NA,
  manual_add_traj1 = NA,
  manual_add_traj2 = NA
)
}
\arguments{
\item{res_mat}{the 9-columned data frame that is the output of \code{segment_genes_along_trajectories}}

\item{nrow1}{the total number of rows in \code{dat1} when used in \code{segment_genes_along_trajectories}}

\item{nrow2}{the total number of rows in \code{dat2} when used in \code{segment_genes_along_trajectories}}

\item{common_n}{the argument \code{common_n} as used in \code{segment_genes_along_trajectories}}

\item{threshold}{the threshold on what qualifies as a gene worth considering. See
\code{eSVD:::.find_common_genes} or \code{eSVD:::.find_trajectory_genes}}

\item{number_of_genes}{maximum number of genes (who's significance exceeds the threshold) that are included
at each position of pseudotime}

\item{manual_add_common}{gene indices (from 1 through \code{max(res_mat$idx)}) (possibly \code{NA})
that are manually included as highly expressed genes shared in common between the two trajectories}

\item{manual_add_traj1}{gene indices (from 1 through \code{max(res_mat$idx)}) (possibly \code{NA})
that are manually included as highly expressed genes in trajectory 1}

\item{manual_add_traj2}{gene indices (from 1 through \code{max(res_mat$idx)}) (possibly \code{NA})
that are manually included as highly expressed genes in trajectory 2}
}
\value{
a list with entries \code{common_genes}, \code{traj1_genes} and \code{traj2_genes},
each containing the index of genes (ranging from \code{1} to \code{max(res_mat$idx)}) in order of
pseudotime
}
\description{
This uses the output of \code{segment_genes_along_trajectories} to find \code{number_of_genes}
highly informative genes along each position of the pseudotime (defined in terms of
indices, as \code{1} through \code{common_n + (nrow(dat1) - commom_n) + (nrow(dat2) - common_n)},
the arguments into \code{segment_genes_along_trajectories}). The function then order such
genes by the midpoint of their significance (as defined in the output of \code{segment_genes_along_trajectories}).
}
