% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_manipulation.R
\name{partition_cells_using_pseudotime}
\alias{partition_cells_using_pseudotime}
\title{Extract which cells are common and which cells are trajectory-specific}
\usage{
partition_cells_using_pseudotime(
  pseudotime_df,
  trajectory_1_clusters,
  trajectory_2_clusters,
  threshold_common_time = NA
)
}
\arguments{
\item{pseudotime_df}{the output from the function \code{eSVD::construct_pseudotime_trajectory_matrix}}

\item{trajectory_1_clusters}{vector containing integers from 1 to \code{max(pseudotime_df$cluster_labels)}}

\item{trajectory_2_clusters}{vector containing integers from 1 to \code{max(pseudotime_df$cluster_labels)}, and
does not have an intersection with \code{trajectory_1_clusters}}

\item{threshold_common_time}{numeric or can be \code{NA}, in which case it uses the maximum pseudotime
for a cell not in either trajectory}
}
\value{
a list of integers
}
\description{
Using the output of \code{eSVD::construct_pseudotime_trajectory_matrix} (called
\code{pseudotime_df}) as well
as which clusters in the \code{cluster_labels} column of \code{pseudotime_df}
correspond to which trajectories, construct 3 vectors of indices (representing different cells) --
one for which cells are common (shared among the two trajectories), and one for each
of the two trajectories. Within each vectors, the output indices are ordered according to the
corresponding cells' pseudotimes.
}
