% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tuning.R
\name{plot_prediction_against_observed}
\alias{plot_prediction_against_observed}
\title{Plotting diagnostic to determine goodness of fit}
\usage{
plot_prediction_against_observed(
  dat,
  nat_mat_list,
  family,
  missing_idx_list = list(1:prod(dim(dat))),
  width = 0.8,
  scalar = NA,
  plot = T,
  max_points = 5e+05,
  ...
)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{nat_mat_list}{list of natural parameter matrices, each of same dimension as \code{dat}}

\item{family}{character such as \code{"gaussian"} or \code{"exponential"}}

\item{missing_idx_list}{list of missing indices, same length as \code{nat_mat_list}}

\item{width}{plotting parameter, controlling quantile}

\item{scalar}{additional parameter needed to compute distribution corresponding to \code{family}}

\item{plot}{boolean}

\item{...}{additional plotting parameters}

\item{seq_max}{plotting parameter, controlling x- and y-axes}
}
\value{
either nothing if \code{plot} is \code{TRUE} (and a plot is shown) or the principle angle otherwise
}
\description{
Plotting diagnostic to determine goodness of fit
}
