% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{prepare_data_for_segmentation}
\alias{prepare_data_for_segmentation}
\title{Prepare data from segmentation downstrem}
\usage{
prepare_data_for_segmentation(
  dat,
  cluster_labels,
  curve_list,
  min_traj_pseudotime = NA,
  cluster_removal_idx_vec = NA,
  cluster_removal_time_vec = NA
)
}
\arguments{
\item{dat}{\code{n} by \code{p} dataset}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{curve_list}{output from \code{eSVD::slingshot}}

\item{min_traj_pseudotime}{positive numeric or \code{NA}}

\item{cluster_removal_idx_vec}{vector of positive integers or a single \code{NA}}

\item{cluster_removal_time_vec}{vector of positive numerics or a single \code{NA}}
}
\value{
a list containing \code{dat1} and \code{dat2} (two datasets formed from a subset of rows from \code{dat})
as well as three vectors of indicies, \code{cell_idx_common}, \code{cell_idx_traj1} and \code{cell_idx_traj2}.
Here, \code{dat1} is formed from \code{dat[c(cell_idx_common,cell_idx_traj1),]}
and \code{dat2} is formed from \code{dat[c(cell_idx_common,cell_idx_traj2),]}.
}
\description{
This function is meant to preprocess \code{dat} so
\code{eSVD::segment_genes_along_trajectories} can be used afterwards.
Here, \code{curve_list} is the \code{slingshot} object returned by
\code{eSVD::slingshot}. This function requires that \code{curve_list} contains exactly two trajectories.
}
\details{
Here, \code{min_traj_pseudotime} is a
pseudotime (according to \code{curve_list}) of which cells that are specific to one of the trajectories
(according to \code{eSVD::construct_pseudotime_trajectory_matrix}) with a smaller pseudotime
\code{min_traj_pseudotime} are removed.

Similarly, \code{cluster_removal_idx_vec} and \code{cluster_removal_time_vec} are vectors of the same
length, of which the cells with a cluster label of the first element in \code{cluster_removal_idx_vec}
with a pseudotime smaller than the first element in \code{cluster_removal_time_vec} are also
removed. (And then second, and so on.)
}
