% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{prepare_trajectory}
\alias{prepare_trajectory}
\title{Prepare trajectories for usage}
\usage{
prepare_trajectory(curve, target_length)
}
\arguments{
\item{curve}{\code{principal_curve} object}

\item{target_length}{target length of the curve.}
}
\value{
numeric matrix
}
\description{
This function only extracts the \code{s} component of the curve, where the curves
are \code{principal_curve} objects, typically returned within the output of \code{eSVD::slingshot}.
This function truncates the curve (starting from the pseudotime index of 0, according
to the \code{ord} component of \code{curve}) so the total length is less that \code{target_length}
}
