% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{screened_genes}
\alias{screened_genes}
\title{Genes used for oligodendrocyte analysis}
\format{
A data frame with 983 rows and one column
\describe{
  \item{gene}{Name of each of the 983 genes}
}
}
\source{
The gene names come from the oligodendrocyte dataset (Marques et al., see
\url{https://github.com/linnykos/esvd/blob/master/data/data.R}), using the screening
procedure described in the analysis file \url{https://github.com/linnykos/esvd/blob/master/main/step0_screening.R}
(The commented out steps are the steps used.) These genes are saved to enable researchers
to more-easily reproduce the eSVD components of the analysis without needing to
rerun the screening procedure again.
}
\description{
Genes used for oligodendrocyte analysis
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
