% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_genes.R
\name{segment_genes_along_trajectories}
\alias{segment_genes_along_trajectories}
\title{Segment the genes along the trajectory}
\usage{
segment_genes_along_trajectories(
  dat1,
  dat2,
  common_n,
  standardize = T,
  ncores = NA,
  verbose = F,
  ...
)
}
\arguments{
\item{dat1}{\code{n1} by \code{p} dataset}

\item{dat2}{\code{n2} by \code{p} dataset}

\item{common_n}{positive integer value denoting that the first \code{common_n} (smaller
than \code{nrow(dat1)} or \code{nrow(dat2)}) rows of
\code{dat1} and \code{dat2} are equivalent}

\item{standardize}{boolean}

\item{ncores}{number of cores}

\item{verbose}{boolean}

\item{...}{additional arguments for \code{eSVD:::.circular_segmentation}}
}
\value{
a list with two elements. The first, called \code{df}
is a data frame with 9 columns and \code{ncol(dat1)} rows, where each row
contains statistics
for each gene across both \code{dat1} and \code{dat2}. The second, called \code{segmentation_fit}
is a list of \code{p} elements, of which is each contains 4 elements:
two elements called \code{cut_1} and \code{cut_2} (containing elements \code{i}, \code{j} and
\code{obj_val}) which describe where the "highly expressed segment" from cell index \code{i} to \code{j}
for either \code{dat1} or \code{dat2} respectively as well as the circular binary segmentation objective
value, and \code{vec1_smooth} and \code{vec2_smooth} (containing the smoothed signal based on
\code{eSVD:::.np_smoother}).
}
\description{
Given two datasets, \code{dat1} and \code{dat2}, whose first \code{common_n} rows
are the same, use the \code{eSVD:::.find_highly_expressed_region} function to output
statistics of each gene in both datasets.
}
\details{
As suggested by the interface, this function can only segment genes when exactly two
trajectories are estimated.
}
