% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{slingshot}
\alias{slingshot}
\title{Use slingshot to estimate the cell development trajectories}
\usage{
slingshot(
  dat,
  cluster_labels,
  starting_cluster,
  cluster_group_list = NA,
  use_initialization = F,
  reduction_percentage = 0.1,
  shrink = 1,
  thresh = 0.001,
  max_iter = 15,
  upscale_factor = NA,
  verbose = F
)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{starting_cluster}{the "origin" cluster that all the lineages will start
from}

\item{cluster_group_list}{list denoting the hierarchy and order of the clusters}

\item{use_initialization}{use principal curves on each cluster to prune the
list of possible neighboring clusters}

\item{reduction_percentage}{numeric to multiplicatively expand or shrink the dataset,
where 1 keeps the dataset unchanged.
For larger values, the dataset is expanded so the estimated trajectories are wigglier,
and for smalelr values, the dataset is shrunk so the estimated trajectories are smoother.}

\item{shrink}{shrinkage factor}

\item{thresh}{parameter to determine convergence}

\item{max_iter}{maximum number of iterations}

\item{upscale_factor}{positive numeric (between 0 and 1) that controls how much to upweight the clusters,
with 1 being (almost) equal cluster sizes and 0 being no upweighting. This does not affect the
estimation of the lineages via \code{.get_lineages}, only the estimation of the curves
given the lineages via \code{.get_curves}.}

\item{verbose}{boolean}
}
\value{
a list containing the lineages under \code{lineages},
the list of curves as \code{principal_curve} objects under
\code{curves} and the clustering matrix under \code{cluster_mat}
}
\description{
Use slingshot to estimate the cell development trajectories
}
