% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{slingshot_3dplot}
\alias{slingshot_3dplot}
\title{3D plotting function}
\usage{
slingshot_3dplot(
  dat,
  cluster_labels,
  bg_col_vec,
  bg_cex = 0.4,
  cluster_center = NA,
  center_cex = 2,
  center_col_vec = NA,
  ...
)
}
\arguments{
\item{dat}{\code{n} by \code{k} (required to be 3 or larger) dataset, of which the first 3 columns
are plotted}

\item{cluster_labels}{numeric of the same length as \code{nrow(dat)}, where the values are consecutive
integers from 1 to \code{max(cluster_labels)}}

\item{bg_col_vec}{vector of colors of the same length as \code{max(cluster_labels)}}

\item{bg_cex}{numeric for plotting, denoting the size of each point (i.e., row) in \code{dat}}

\item{cluster_center}{location of the cluster centers, possibly \code{NA}. Otherwise, it is a
numeric matrix of \code{max(cluster_labels)} rows and at least 3 columns.}

\item{center_cex}{numeric for plotting, denoting the size of each cluster center in \code{cluster_center}}

\item{center_col_vec}{vector of colors of the same length as \code{max(cluster_labels)}}

\item{...}{additional plotting parameter}
}
\value{
no return. A plot is produced
}
\description{
Relies heavily on the \code{plot3D} package
}
