% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tuning.R
\name{tuning}
\alias{tuning}
\title{Tuning distributions}
\usage{
tuning(dat, u_mat, v_mat, family, r_min = 1, r_max = 100)
}
\arguments{
\item{dat}{data matrix}

\item{u_mat}{initial fit from \code{fit_factorization}}

\item{v_mat}{initial fit from \code{fit_factorization}}

\item{family}{character such as \code{"gaussian"} or \code{"exponential"}}

\item{r_min}{smallest value}

\item{r_max}{largest value}
}
\value{
value
}
\description{
Tuning distributions
}
