% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tuning.R
\name{tuning_scalar}
\alias{tuning_scalar}
\title{Tuning negative binomial or curved Gaussian scalar}
\usage{
tuning_scalar(
  dat,
  family,
  iter_max = 5,
  search_min = 1,
  search_max = 2000,
  verbose = F,
  ...
)
}
\arguments{
\item{dat}{\code{n} by \code{p} data matrix}

\item{family}{character (either \code{"neg_binom"} or \code{"curved_gaussian"})}

\item{iter_max}{positive intger}

\item{search_min}{positive intger}

\item{search_max}{positive intger, larger than \code{search_min}}

\item{verbose}{boolean}

\item{...}{all remaining parameters needed for either \code{initialization} or \code{matrix_factorization}}
}
\value{
vector of length \code{search_iter}
}
\description{
Tuning negative binomial or curved Gaussian scalar
}
