% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tuning.R
\name{tuning_select_scalar}
\alias{tuning_select_scalar}
\title{Select tuning parameter for a bunch of matrix completed fits}
\usage{
tuning_select_scalar(
  dat,
  nat_mat_list_list,
  family,
  missing_idx_list = list(1:prod(dim(dat))),
  width = 0.8,
  scalar_vec = rep(NA, length(nat_mat_list_list)),
  compute_percentage = F
)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{nat_mat_list_list}{list of lists of natural parameter matrices, each of same dimension as \code{dat}.
The length of the list \code{nat_mat_list_list} is equal to \code{length{scalar_vec}},
the number of different scalars to select from, and is
comprised of individual lists. The length of each of such lists within each element of
\code{nat_mat_list_list} is equal to the length of \code{length(missing_idx_list)}, the
number of different trials for each parameter setting}

\item{family}{character such as \code{"gaussian"}, \code{"exponential"}, \code{"neg_binom"} or \code{"curved_gaussian"}}

\item{missing_idx_list}{list of missing indices, same length as \code{nat_mat_list_list[[1]]}}

\item{width}{parameter, controlling quantile of prediction region}

\item{scalar_vec}{vector of additional parameters needed to compute distribution corresponding to \code{family},
of length equal to \code{length(nat_mat_list_list)}}

\item{compute_percentage}{boolean to whether or not compute the percentage of points that fall within the \code{width}-sized
prediction region. It is suggested to keep this as \code{FALSE} since this might be slow to compute.}
}
\value{
a list
}
\description{
Select tuning parameter for a bunch of matrix completed fits
}
