% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_tube.R
\name{bootstrap_curves}
\alias{bootstrap_curves}
\title{Bootstrap the curves via resampleing of each cluster}
\usage{
bootstrap_curves(dat, cluster_labels, starting_cluster, cluster_group_list,
  trials = 100, cores = NA, ...)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{starting_cluster}{the "origin" cluster that all the lineages will start
from}

\item{cluster_group_list}{list denoting the hierarchy and order of the clusters}

\item{trials}{numeric}

\item{cores}{numeric}

\item{...}{additional parameters for \code{slingshot}}
}
\value{
a list
}
\description{
Bootstrap the curves via resampleing of each cluster
}
