% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_tube.R
\name{compute_curve_sd}
\alias{compute_curve_sd}
\title{Compute the bootstrap uncertainty}
\usage{
compute_curve_sd(
  target_curve_list,
  bootstrap_curve_list,
  quantile_inner = 0.95,
  quantile_outer = 0.95,
  ncores = NA,
  verbose = F
)
}
\arguments{
\item{target_curve_list}{a list (for example, the \code{curves} output from the \code{slingshot} function)}

\item{bootstrap_curve_list}{a list
(for example, the output from the \code{bootstrap_curves} function)}

\item{quantile_inner}{quantile value representing, for each point on the estimated curve,
how far do you want to encapsulate all the bootstrapped curves}

\item{quantile_outer}{quantile value representing, among all the points on the estimated curve,
how far do you want to encapuslate all their values dictated by quantile_inner}

\item{ncores}{numeric}

\item{verbose}{boolean}
}
\value{
a numeric
}
\description{
Try to compute the "radius" of uncertainty by conditioning on the same curves
}
