% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{.reparameterize}
\alias{.reparameterize}
\title{Function to reparameterize two matrices}
\usage{
.reparameterize(u_mat, v_mat, equal_covariance = F)
}
\arguments{
\item{u_mat}{matrix of dimension \code{n} by \code{k}}

\item{v_mat}{matrix of dimension \code{p} by \code{k}}

\item{equal_covariance}{boolean}
}
\value{
list of two matrices
}
\description{
test
}
\details{
Designed to output matrices of the same dimension as \code{u_mat}
and \code{v_mat}, but linearly transformed so \code{u_mat \%*\% t(v_mat)}
is preserved but either \code{u_mat \%*\% t(u_mat)} is diagonal and equal to
\code{v_mat \%*\% t(v_mat)} (if \code{equal_covariance} is \code{FALSE})
or \code{u_mat \%*\% t(u_mat)/nrow(u_mat)} is diagonal and equal to
\code{v_mat \%*\% t(v_mat)/nrow(v_mat)} (if \code{equal_covariance} is \code{TRUE})
}
