% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tuning.R
\name{plot_prediction_against_observed}
\alias{plot_prediction_against_observed}
\title{Plotting diagnostic to determine goodness of fit}
\usage{
plot_prediction_against_observed(
  dat,
  nat_mat_list,
  family,
  missing_idx_list = list(1:prod(dim(dat))),
  width = 0.8,
  scalar = NA,
  plot = T,
  max_points = 5e+05,
  tol = 0.95,
  xlim = NA,
  ylim = NA,
  transparency = 0.2,
  cex_text = 1,
  ...
)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{nat_mat_list}{list of natural parameter matrices, each of same dimension as \code{dat}}

\item{family}{character such as \code{"gaussian"}, \code{"exponential"}, \code{"neg_binom"} or \code{"curved_gaussian"}}

\item{missing_idx_list}{list of missing indices, same length as \code{nat_mat_list}}

\item{width}{parameter, controlling quantile of prediction region}

\item{scalar}{additional parameter needed to compute distribution corresponding to \code{family}}

\item{plot}{boolean}

\item{max_points}{maximum number of points to be shown in the scatterplot, purely for visualization purposes only}

\item{tol}{parameter between \code{0} and \code{1} for how strict (\code{1} being the strictest) to measure
if the principal angle falls within the prediction region}

\item{xlim}{plotting parameter}

\item{ylim}{plotting parameter}

\item{transparency}{plotting parameter}

\item{cex_text}{plotting parameter}

\item{...}{additional plotting parameters}
}
\value{
either nothing if \code{plot} is \code{TRUE} (and a plot is shown) or the principle angle otherwise
}
\description{
Plotting diagnostic to determine goodness of fit
}
